+++ api-derive/staking/ownExposure.ts
@@ -1,6 +1,9 @@
 
 import type { Observable } from 'https://esm.sh/rxjs@7.8.1';
+import type { Option, u32 } from 'https://deno.land/x/polkadot/types/mod.ts';
 import type { EraIndex } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { SpStakingExposurePage } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { AnyNumber } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 import type { DeriveApi, DeriveOwnExposure } from '../types.ts';
 
 import { combineLatest, map, of } from 'https://esm.sh/rxjs@7.8.1';
@@ -8,24 +11,28 @@ import { combineLatest, map, of } from 'https://esm.sh/rxjs@7.8.1';
 import { firstMemo, memo } from '../util/index.ts';
 import { erasHistoricApplyAccount } from './util.ts';
 
-export function _ownExposures (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive: boolean) => Observable<DeriveOwnExposure[]> {
-  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive: boolean): Observable<DeriveOwnExposure[]> =>
-    eras.length
+export function _ownExposures (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive: boolean, page: u32 | AnyNumber) => Observable<DeriveOwnExposure[]> {
+  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive: boolean, page: u32 | AnyNumber): Observable<DeriveOwnExposure[]> => {
+    return eras.length
       ? combineLatest([
+        // Backwards and forward compat for historical integrity when using `erasHistoricApplyAccount`
         combineLatest(eras.map((e) => api.query.staking.erasStakersClipped(e, accountId))),
-        combineLatest(eras.map((e) => api.query.staking.erasStakers(e, accountId)))
+        combineLatest(eras.map((e) => api.query.staking.erasStakers(e, accountId))),
+        combineLatest(eras.map((e) => api.query.staking.erasStakersPaged<Option<SpStakingExposurePage>>(e, accountId, page))),
+        combineLatest(eras.map((e) => api.query.staking.erasStakersOverview(e, accountId)))
       ]).pipe(
-        map(([clp, exp]): DeriveOwnExposure[] =>
-          eras.map((era, index) => ({ clipped: clp[index], era, exposure: exp[index] }))
+        map(([clp, exp, paged, expMeta]): DeriveOwnExposure[] =>
+          eras.map((era, index) => ({ clipped: clp[index], era, exposure: exp[index], exposureMeta: expMeta[index], exposurePaged: paged[index] }))
         )
       )
-      : of([])
+      : of([]);
+  }
   );
 }
 
 export const ownExposure = /*#__PURE__*/ firstMemo(
-  (api: DeriveApi, accountId: Uint8Array | string, era: EraIndex) =>
-    api.derive.staking._ownExposures(accountId, [era], true)
+  (api: DeriveApi, accountId: Uint8Array | string, era: EraIndex, page?: u32 | AnyNumber) =>
+    api.derive.staking._ownExposures(accountId, [era], true, page || 0)
 );
 
 export const ownExposures = /*#__PURE__*/ erasHistoricApplyAccount('_ownExposures');